Fixed issue with --width parameter
KernelDepth and EnvelopeDepth were set from the default value of --depth instead of the calculated value. So these initializations were moved to after the --depth calculation.
Now the resulting mesh with --depth calculated from width will be identical to the mesh calculated when --depth is set to the same value directly, given they are derived from the same data.